* MonoTorrent/MonoTorrent.Client/Managers/ConnectionManager.cs: Fix   a math overflow error. Thanks to nivanov.
svn path=/trunk/bitsharp/; revision=153792

+++ src/MonoTorrent/MonoTorrent.Client/Managers/ConnectionManager.cs
@@ -605,7 +605,7 @@ namespace MonoTorrent.Client
                     PieceMessage pm = (PieceMessage)msg;
                     pm.Data = BufferManager.EmptyBuffer;
                     ClientEngine.BufferManager.GetBuffer(ref pm.Data, pm.ByteLength);
-                    engine.DiskManager.QueueRead(id.TorrentManager, pm.StartOffset + (pm.PieceIndex * id.TorrentManager.Torrent.PieceLength), pm.Data, pm.RequestLength, delegate
+                    engine.DiskManager.QueueRead(id.TorrentManager, pm.StartOffset + ((long) pm.PieceIndex * id.TorrentManager.Torrent.PieceLength), pm.Data, pm.RequestLength, delegate
                     {
                         handle.Set();
                     });
